SalaryCalculationPercentage and ReusingVariables

+++ src/SalaryCalculationPercentage.java
@@ -1,9 +1,9 @@
 public class SalaryCalculationPercentage {
     public static void main(String[] args) {
         // Create a class that calculates a certain percentage of a given salary.
-        float salary = 2500.00F;
-        float percentage = 30;
-        float percentageSalary = salary * percentage / 100;
+        double salary = 2500.00;
+        double percentage = 30;
+        double percentageSalary = salary * percentage / 100;
         System.out.println("The percentage calculated was "+percentageSalary);
     }
 }
